Move Agent class to its own file
